refactor(api): update api routes
